docs: update contribution guidelines
Changed the "Getting Started" section title to "How To Contribute?"
because there is no following workflow. Removed development workflow
in the `CONTRIBUTING.md` file and replaced with a reference to the
"Development Workflow" section in the `README.md` file. Updated
the "Commit Guidelines" to link to the official Conventional
Commits website. Fixed typo for named-link list in the
`CONTRIBUTING.md` file.